expose MonitorServiceSample
