+++ include/fastdds_statistics_backend/types/types.hpp
@@ -23,6 +23,7 @@
 #include <fastdds_statistics_backend/nlohmann-json/json.hpp>
 #include <fastdds_statistics_backend/types/Bitmask.hpp>
 #include <fastdds_statistics_backend/types/EntityId.hpp>
+#include <fastdds_statistics_backend/topic_types/monitorservice_types.h>
 
 #include <chrono>
 
@@ -353,6 +354,291 @@ enum class StatisticKind
     SUM
 };
 
+
+/*
+ * Base class for all monitor service samples. It adds the timepoint for the sample
+ */
+struct MonitorServiceSample
+{
+    MonitorServiceSample(
+            StatusKind sample_kind = StatusKind::INVALID,
+            EntityStatus sample_status = EntityStatus::OK)
+        : kind(sample_kind), status(sample_status)
+    {
+    }
+
+    virtual ~MonitorServiceSample() = default;
+
+    virtual void clear();
+
+    inline bool operator ==(
+            const MonitorServiceSample& other) const noexcept
+    {
+        return (kind == other.kind && status == other.status && src_ts == other.src_ts);
+    }
+
+    inline bool operator !=(
+            const MonitorServiceSample& other) const noexcept
+    {
+        return !(*this == other);
+    }
+
+    StatusKind kind;
+    EntityStatus status;
+    std::chrono::system_clock::time_point src_ts;
+};
+
+
+///*
+// * Proxy data of the status data
+// */
+//struct ProxySample : MonitorServiceSample
+//{
+//    ProxySample(): MonitorServiceSample(StatusKind::PROXY)
+//    {
+//    }
+//
+//    virtual ~ProxySample() = default;
+//
+//    void clear() final;
+//  
+//    inline bool operator ==(
+//            const ProxySample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && entity_proxy == other.entity_proxy);
+//    }
+//
+//    inline bool operator !=(
+//            const ProxySample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    std::vector<uint8_t> entity_proxy;
+//};
+///*
+// * Connection list status data
+// */
+//struct ConnectionListSample : MonitorServiceSample
+//{
+//    ConnectionListSample(): MonitorServiceSample(StatusKind::CONNECTION_LIST)
+//    {
+//    }
+//
+//    virtual ~ConnectionListSample() = default;
+//
+//    void clear() final;
+//    
+//    inline bool operator ==(
+//            const ConnectionListSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && connection_list == other.connection_list);
+//    }
+//
+//    inline bool operator !=(
+//            const ConnectionListSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    std::vector<Connection> connection_list;
+//};
+//
+/*
+ * Incompatible Qos status data
+ */
+struct IncompatibleQosSample : MonitorServiceSample
+{
+    IncompatibleQosSample(): MonitorServiceSample(StatusKind::INCOMPATIBLE_QOS)
+    {
+    }
+
+    virtual ~IncompatibleQosSample() = default;
+
+    void clear() final;
+
+    inline bool operator ==(
+            const IncompatibleQosSample& other) const noexcept
+    {
+        return (MonitorServiceSample::operator ==(other) && incompatible_qos_status == other.incompatible_qos_status);
+    }
+
+    inline bool operator !=(
+            const IncompatibleQosSample& other) const noexcept
+    {
+        return !(*this == other);
+    }
+
+    eprosima::fastdds::statistics::IncompatibleQoSStatus_s incompatible_qos_status;
+};
+//
+///*
+// * Inconsistent topic status data
+// */
+//struct InconsistentTopicSample : MonitorServiceSample
+//{
+//    InconsistentTopicSample(): MonitorServiceSample(StatusKind::INCONSISTENT_TOPIC)
+//    {
+//    }
+//
+//    virtual ~InconsistentTopicSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const InconsistentTopicSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && inconsistent_topic_status == other.inconsistent_topic_status);
+//    }
+//
+//    inline bool operator !=(
+//            const InconsistentTopicSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    InconsistentTopicStatus_s inconsistent_topic_status;
+//};
+//
+///*
+// * Liveliness lost status data
+// */
+//struct LivelinessLostSample : MonitorServiceSample
+//{
+//    LivelinessLostSample(): MonitorServiceSample(StatusKind::LIVELINESS_LOST)
+//    {
+//    }
+//
+//    virtual ~LivelinessLostSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const LivelinessLostSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && liveliness_lost_status == other.liveliness_lost_status);
+//    }
+//
+//    inline bool operator !=(
+//            const LivelinessLostSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    LivelinessLostStatus_s liveliness_lost_status;
+//};
+//
+///*
+// * Liveliness changed status data
+// */
+//struct LivelinessChangedSample : MonitorServiceSample
+//{
+//    LivelinessChangedSample(): MonitorServiceSample(StatusKind::LIVELINESS_CHANGED)
+//    {
+//    }
+//
+//    virtual ~LivelinessChangedSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const LivelinessChangedSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && liveliness_lost_status == other.liveliness_lost_status);
+//    }
+//
+//    inline bool operator !=(
+//            const LivelinessChangedSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    LivelinessChangedStatus_s liveliness_changed_status;
+//};
+//
+///*
+// * Deadline missed status data
+// */
+//struct DeadlineMissedSample : MonitorServiceSample
+//{
+//    DeadlineMissedSample(): MonitorServiceSample(StatusKind::DEADLINE_MISSED)
+//    {
+//    }
+//
+//    virtual ~DeadlineMissedSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const DeadlineMissedSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && deadline_missed_status == other.deadline_missed_status);
+//    }
+//
+//    inline bool operator !=(
+//            const DeadlineMissedSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    DeadlineMissedStatus_s deadline_missed_status;
+//};
+//
+///*
+// * Sample lost status data
+// */
+//struct SampleLostSample : MonitorServiceSample
+//{
+//    SampleLostSample(): MonitorServiceSample(StatusKind::SAMPLE_LOST)
+//    {
+//    }
+//
+//    virtual ~SampleLostSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const SampleLostSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && sample_lost_status == other.sample_lost_status);
+//    }
+//
+//    inline bool operator !=(
+//            const SampleLostSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    SampleLostStatus_s sample_lost_status;
+//};
+//
+///*
+// * Statuses size status data
+// */
+//struct StatusesSizeSample : MonitorServiceSample
+//{
+//    StatusesSizeSample(): MonitorServiceSample(StatusKind::STATUSES_SIZE)
+//    {
+//    }
+//
+//    virtual ~StatusesSizeSample() = default;
+//
+//    void clear() final;
+//
+//    inline bool operator ==(
+//            const StatusesSizeSample& other) const noexcept
+//    {
+//        return (MonitorServiceSample::operator ==(other) && statuses_size == other.statuses_size);
+//    }
+//
+//    inline bool operator !=(
+//            const StatusesSizeSample& other) const noexcept
+//    {
+//        return !(*this == other);
+//    }
+//
+//    octet statuses_size;
+//};
 
 } //namespace statistics_backend
 } //namespace eprosima
