+++ src/cpp/StatisticsBackendData.hpp
@@ -33,8 +33,6 @@
 #include <fastdds_statistics_backend/types/types.hpp>
 #include <fastdds_statistics_backend/types/EntityId.hpp>
 
-#include <topic_types/monitorservice_types.h>
-
 #include <database/database.hpp>
 #include "Monitor.hpp"
