+++ src/cpp/subscriber/StatisticsReaderListener.cpp
@@ -26,7 +26,6 @@
 #include <fastrtps/types/TypesBase.h>
 
 #include <database/database_queue.hpp>
-#include <topic_types/monitorservice_types.h>
 
 namespace eprosima {
 namespace statistics_backend {
